requirements.txt
# NEST-DRUG: Nested-Learning Platform for Drug Discovery
# Requirements file

# Core ML
torch>=2.0.0
torch-geometric>=2.4.0
pytorch-lightning>=2.0.0

# Molecular processing
rdkit>=2023.9.1
deepchem>=2.7.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Drug discovery datasets
PyTDC>=0.4.0
chembl_webresource_client>=0.10.8

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
umap-learn>=0.5.3

# Utilities
tqdm>=4.65.0
requests>=2.31.0
h5py>=3.8.0
pyarrow>=12.0.0
